swarm - Refactor: fixed agents
