+++ pkgs/swarmauri/swarmauri/agent_factories/__init__.py
@@ -1,0 +1,1 @@
+from swarmauri.agent_factories.concrete import *
