+++ pkgs/swarmauri/swarmauri/agents/__init__.py
@@ -1,2 +1,3 @@
 # -*- coding: utf-8 -*-
 
+from swarmauri.agents.concrete import *
