+++ pkgs/swarmauri/swarmauri/agents/concrete/__init__.py
@@ -1,0 +1,11 @@
+from swarmauri.agents.concrete.SimpleConversationAgent import SimpleConversationAgent
+from swarmauri.agents.concrete.QAAgent import QAAgent
+from swarmauri.agents.concrete.RagAgent import RagAgent
+from swarmauri.agents.concrete.ToolAgent import ToolAgent
+
+__all__ = [
+    "SimpleConversationAgent",
+    "QAAgent",
+    "RagAgent",
+    "ToolAgent",
+]
