+++ pkgs/swarmauri/tests/unit/agents/ToolAgent_unit_test.py
@@ -4,7 +4,7 @@ from swarmauri.llms.concrete.GroqToolModel import GroqToolModel
 from swarmauri.conversations.concrete.Conversation import Conversation
 from swarmauri.tools.concrete.AdditionTool import AdditionTool
 from swarmauri.toolkits.concrete.Toolkit import Toolkit
-from swarmauri.agents.concrete.ToolAgent import ToolAgent
+from swarmauri.agents.concrete import ToolAgent
 from dotenv import load_dotenv
 
 load_dotenv()
